+++ db-queue-scheduler-core/src/main/java/ru/yoomoney/tech/dbqueue/scheduler/internal/queue/ScheduledTaskQueue.java
@@ -59,7 +59,7 @@ public class ScheduledTaskQueue {
     }
 
     private void doInitTask() {
-        if (!scheduledQueueDao.isQueueEmpty(queueConfig.getLocation().getQueueId())) {
+        if (scheduledQueueDao.findQueueTask(queueConfig.getLocation().getQueueId()).isPresent()) {
             log.debug("scheduled task already enqueued: taskDefinition={}", taskDefinition);
             return;
         }