+++ db-queue-scheduler-spring/src/test/java/ru/yoomoney/tech/dbqueue/scheduler/SchedulerTest.java
@@ -6,9 +6,11 @@ import ru.yoomoney.tech.dbqueue.scheduler.models.ScheduledTask;
 import ru.yoomoney.tech.dbqueue.scheduler.models.ScheduledTaskExecutionResult;
 import ru.yoomoney.tech.dbqueue.scheduler.models.SimpleScheduledTask;
 import ru.yoomoney.tech.dbqueue.scheduler.models.info.ScheduledTaskInfo;
+import ru.yoomoney.tech.dbqueue.scheduler.settings.FailureSettings;
 import ru.yoomoney.tech.dbqueue.scheduler.settings.ScheduleSettings;
 import ru.yoomoney.tech.dbqueue.scheduler.settings.ScheduledTaskSettings;
 
+import java.sql.Timestamp;
 import java.time.Duration;
 import java.time.Instant;
 import java.util.Arrays;
@@ -24,6 +26,7 @@ import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.notNullValue;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 /**
  * @author Petr Zinin pgzinin@yoomoney.ru
@@ -54,8 +57,8 @@ public class SchedulerTest extends BaseTest {
         scheduler.schedule(
                 scheduledTask,
                 ScheduledTaskSettings.builder()
-                        .withMaxExecutionLockInterval(Duration.ofHours(1L))
                         .withScheduleSettings(ScheduleSettings.fixedDelay(Duration.ofSeconds(0L)))
+                        .withFailureSettings(FailureSettings.linearBackoff(Duration.ofHours(1L)))
                         .build()
         );
 
@@ -87,14 +90,14 @@ public class SchedulerTest extends BaseTest {
         scheduler1.schedule(
                 scheduledTask,
                 ScheduledTaskSettings.builder()
-                        .withMaxExecutionLockInterval(Duration.ofHours(1L))
+                        .withFailureSettings(FailureSettings.linearBackoff(Duration.ofHours(1L)))
                         .withScheduleSettings(ScheduleSettings.fixedDelay(Duration.ofSeconds(1L)))
                         .build()
         );
         scheduler2.schedule(
                 scheduledTask,
                 ScheduledTaskSettings.builder()
-                        .withMaxExecutionLockInterval(Duration.ofHours(1L))
+                        .withFailureSettings(FailureSettings.linearBackoff(Duration.ofHours(1L)))
                         .withScheduleSettings(ScheduleSettings.fixedDelay(Duration.ofSeconds(1L)))
                         .build()
         );
@@ -149,7 +152,7 @@ public class SchedulerTest extends BaseTest {
         scheduler.schedule(
                 scheduledTask,
                 ScheduledTaskSettings.builder()
-                        .withMaxExecutionLockInterval(Duration.ofHours(1L))
+                        .withFailureSettings(FailureSettings.linearBackoff(Duration.ofHours(1L)))
                         .withScheduleSettings(ScheduleSettings.fixedDelay(Duration.ofSeconds(1L)))
                         .build()
         );
@@ -158,7 +161,7 @@ public class SchedulerTest extends BaseTest {
         assertThrows(RuntimeException.class, () -> scheduler.schedule(
                 scheduledTask,
                 ScheduledTaskSettings.builder()
-                        .withMaxExecutionLockInterval(Duration.ofHours(1L))
+                        .withFailureSettings(FailureSettings.linearBackoff(Duration.ofHours(1L)))
                         .withScheduleSettings(ScheduleSettings.fixedDelay(Duration.ofSeconds(1L)))
                         .build()));
     }
@@ -182,7 +185,7 @@ public class SchedulerTest extends BaseTest {
         scheduler.schedule(
                 scheduledTask,
                 ScheduledTaskSettings.builder()
-                        .withMaxExecutionLockInterval(Duration.ofHours(1L))
+                        .withFailureSettings(FailureSettings.linearBackoff(Duration.ofHours(1L)))
                         .withScheduleSettings(ScheduleSettings.fixedDelay(Duration.ofSeconds(1L)))
                         .build()
         );
@@ -219,7 +222,7 @@ public class SchedulerTest extends BaseTest {
         scheduler.schedule(
                 scheduledTask,
                 ScheduledTaskSettings.builder()
-                        .withMaxExecutionLockInterval(Duration.ofHours(1L))
+                        .withFailureSettings(FailureSettings.linearBackoff(Duration.ofHours(1L)))
                         .withScheduleSettings(ScheduleSettings.fixedDelay(Duration.ofDays(1L)))
                         .build()
         );
@@ -227,5 +230,43 @@ public class SchedulerTest extends BaseTest {
 
         // then
         await().atMost(Duration.ofSeconds(5L)).until(executed::get);
+    }
+
+    @Test
+    public void should_retry_failed_task() {
+        // given
+        Scheduler scheduler = new SpringSchedulerConfigurator()
+                .withDatabaseDialect(DatabaseDialect.POSTGRESQL)
+                .withTableName("scheduled_tasks")
+                .withJdbcOperations(jdbcTemplate)
+                .withTransactionOperations(transactionTemplate)
+                .configure();
+        AtomicBoolean executed = new AtomicBoolean(false);
+        ScheduledTask scheduledTask = SimpleScheduledTask.create(
+                "scheduled-task" + uniqueCounter.incrementAndGet(),
+                () -> {
+                    executed.set(true);
+                    return ScheduledTaskExecutionResult.error();
+                }
+        );
+
+        // when
+        scheduler.start();
+        scheduler.schedule(
+                scheduledTask,
+                ScheduledTaskSettings.builder()
+                        .withFailureSettings(FailureSettings.linearBackoff(Duration.ofMinutes(1L)))
+                        .withScheduleSettings(ScheduleSettings.fixedDelay(Duration.ofHours(1L)))
+                        .build()
+        );
+        scheduler.reschedule(scheduledTask.getIdentity(), Instant.now());
+
+        // then
+        await().atMost(Duration.ofSeconds(5L)).until(executed::get);
+
+        Instant nextExecutionDate = jdbcTemplate.queryForObject("select next_process_at from scheduled_tasks where queue_name = ?",
+                Timestamp.class, scheduledTask.getIdentity().asString()).toInstant();
+        assertTrue(nextExecutionDate.minus(Duration.ofMinutes(1L)).isBefore(Instant.now().plus(Duration.ofMinutes(1L))));
+        assertTrue(nextExecutionDate.plus(Duration.ofMinutes(1L)).isAfter(Instant.now().plus(Duration.ofMinutes(1L))));
     }
 }